fix: Always recreate systemd service file with expanded variables
The previous version only created the service file if it didn't exist,
which meant the old broken configuration (with literal $USER text) was
being reused.

Now we always recreate the service file on each deployment to ensure
it has the latest configuration with properly expanded shell variables.

This will fix the "status 217/USER" error where systemd couldn't
find the user.